requirements.txt
APScheduler==3.11.0
blinker==1.9.0
cachelib==0.13.0
click==8.2.1
Flask==3.1.1
Flask-Compress==1.15
Flask-Session==0.8.0
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
msgspec==0.19.0
packaging==25.0
pillow==11.3.0
tzlocal==5.3.1
Werkzeug==3.1.3
csscompressor==0.9.5
jsmin==3.0.1
